Reverted a little part of pull 293, waiting for a proper printf fix

+++ src/interfaces/text/ec_text.c
@@ -540,9 +540,9 @@ static void text_run_filter(void) {
  */
 static void text_stats(void)
 {
-   DEBUG_MSG("text_stats (pcap) : %llu %llu %llu", (long long unsigned int)GBL_STATS->ps_recv, 
-                                                (long long unsigned int)GBL_STATS->ps_drop,
-                                                (long long unsigned int)GBL_STATS->ps_ifdrop);
+   DEBUG_MSG("text_stats (pcap) : %llu %llu %llu", GBL_STATS->ps_recv, 
+                                                GBL_STATS->ps_drop,
+                                                GBL_STATS->ps_ifdrop);
    DEBUG_MSG("text_stats (BH) : [%lu][%lu] p/s -- [%lu][%lu] b/s", 
          (unsigned long)GBL_STATS->bh.rate_adv, (unsigned long)GBL_STATS->bh.rate_worst, 
          (unsigned long)GBL_STATS->bh.thru_adv, (unsigned long)GBL_STATS->bh.thru_worst); 
@@ -554,18 +554,18 @@ static void text_stats(void)
    DEBUG_MSG("text_stats (queue) : %lu %lu", (unsigned long)GBL_STATS->queue_curr, (unsigned long)GBL_STATS->queue_max); 
   
    
-   fprintf(stdout, "\n Received packets    : %8lld\n", (long long unsigned int)GBL_STATS->ps_recv);
-   fprintf(stdout,   " Dropped packets     : %8lld  %.2f %%\n", (long long unsigned int)GBL_STATS->ps_drop,
+   fprintf(stdout, "\n Received packets    : %8lld\n", GBL_STATS->ps_recv);
+   fprintf(stdout,   " Dropped packets     : %8lld  %.2f %%\n", GBL_STATS->ps_drop,
          (GBL_STATS->ps_recv) ? (float)GBL_STATS->ps_drop * 100 / GBL_STATS->ps_recv : 0 );
-   fprintf(stdout,   " Forwarded           : %8lld  bytes: %8lld\n\n", (long long unsigned int)GBL_STATS->ps_sent, (long long unsigned int)GBL_STATS->bs_sent);
+   fprintf(stdout,   " Forwarded           : %8lld  bytes: %8lld\n\n", GBL_STATS->ps_sent, GBL_STATS->bs_sent);
    
    fprintf(stdout,   " Current queue len   : %lu/%lu\n", (unsigned long)GBL_STATS->queue_curr, (unsigned long)GBL_STATS->queue_max);
    fprintf(stdout,   " Sampling rate       : %d\n\n", GBL_CONF->sampling_rate);
    
    fprintf(stdout,   " Bottom Half received packet : pck: %8lld  byte: %8lld\n", 
-         (long long unsigned int)GBL_STATS->bh.pck_recv, (long long unsigned int)GBL_STATS->bh.pck_size);
+         GBL_STATS->bh.pck_recv, GBL_STATS->bh.pck_size);
    fprintf(stdout,   " Top Half received packet    : pck: %8lld  byte: %8lld\n", 
-         (long long unsigned int)GBL_STATS->th.pck_recv, (long long unsigned int)GBL_STATS->th.pck_size);
+         GBL_STATS->th.pck_recv, GBL_STATS->th.pck_size);
    fprintf(stdout,   " Interesting packets         : %.2f %%\n\n",
          (GBL_STATS->bh.pck_recv) ? (float)GBL_STATS->th.pck_recv * 100 / GBL_STATS->bh.pck_recv : 0 );
